+++ programs/dcarbon-contract/src/instructions/config/set_rate.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 
 use crate::instructions::SetConfig;
 
-pub fn set_rate(ctx: Context<SetConfig>, rate: u64) -> Result<()> {
+pub fn set_rate(ctx: Context<SetConfig>, rate: f64) -> Result<()> {
     let contract_config = &mut ctx.accounts.contract_config;
 
     contract_config.rate = rate;
