+++ programs/dcarbon-contract/src/state/config.rs
@@ -10,8 +10,8 @@ use crate::state::{BpfWriter, DeviceLimit};
 #[derive(Debug, InitSpace)]
 pub struct ContractConfig {
     // unlock rate for DCARBON when mint sFT
-    pub rate: u64,
-    pub minting_fee: u64,
+    pub rate: f64,
+    pub minting_fee: f64,
     pub mint: Pubkey,
     #[max_len(100)]
     pub minting_limits: Vec<DeviceLimit>,
@@ -22,7 +22,7 @@ impl ContractConfig {
     pub const PREFIX_SEED: &'static [u8] = b"contract_config";
 
     pub fn validate(&self, config_args: ConfigArgs) -> Result<()> {
-        if config_args.rate <= 0 || config_args.minting_fee <= 0 || config_args.governance_amount <= 0 {
+        if config_args.rate <= 0.0 || config_args.minting_fee <= 0.0 || config_args.governance_amount <= 0 {
             return Err(DCarbonError::InitArgsInvalid.into());
         }
         Ok(())
